+++ src/main/core/Resources/config/services/subscriber.yml
@@ -88,7 +88,6 @@ services:
             - '@Claroline\CoreBundle\API\Serializer\ParametersSerializer'
             - '@Claroline\CoreBundle\Manager\LocaleManager'
             - '@Claroline\ThemeBundle\Manager\IconSetManager'
-
     # Tools
     Claroline\CoreBundle\Subscriber\Tool\LocationsSubscriber:
         tags: [kernel.event_subscriber]
@@ -108,3 +107,7 @@ services:
         tags: [kernel.event_subscriber]
         arguments:
             - '@security.authorization_checker'
+
+    Claroline\CoreBundle\Subscriber\Tool\ConnectionMessagesSubscriber:
+      tags: [ kernel.event_subscriber ]
+
