+++ src/main/core/Resources/modules/administration/connection-messages/components/tool.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+import {ToolPage} from '#/main/core/tool/containers/page'
+
+const ConnectionMessagesTool = (props) =>
+    <ToolPage>
+        Hello World
+    </ToolPage>
+
+export { ConnectionMessagesTool }
