+++ src/main/core/Resources/modules/administration/connection-messages/index.js
@@ -1,0 +1,6 @@
+import {ConnectionMessagesTool} from '#/main/core/administration/connection-messages/components/tool'
+import {reducer} from '#/main/core/administration/connection-messages/store'
+export default {
+  component: ConnectionMessagesTool,
+  store: reducer
+}
