+++ src/main/core/Resources/modules/administration/connection-messages/store/index.js
@@ -1,0 +1,4 @@
+import {reducer} from '#/main/core/administration/connection-messages/store/reducer'
+import {selectors} from '#/main/core/administration/connection-messages/store/selector'
+
+export {reducer, selectors}
