+++ src/main/core/Resources/translations/tools.en.json
@@ -7,6 +7,7 @@
     "parameters": "Settings",
     "resources": "Resources",
     "community": "Community",
+    "connection_messages": "Messages at connection",
     "main_settings": "Settings",
     "technical_settings": "Technical settings",
     "appearance_settings": "Appearance",
