+++ src/main/core/Resources/translations/tools.fr.json
@@ -7,6 +7,7 @@
     "parameters": "Paramètres",
     "resources": "Ressources",
     "community": "Communauté",
+    "connection_messages": "Messages à la connexion",
     "main_settings": "Paramètres",
     "technical_settings": "Paramètres techniques",
     "appearance_settings": "Apparence",
